frontend: add custom auth hook
